asistente-asignatura/scriptsAuxiliares/conversores/markdown2Pdf/pyproject.toml
[tool.poetry]
name = "markdown2pdf-pandoc"
version = "2.0.0"
description = "Conversor Markdown multi-formato usando Pandoc"
authors = ["Teaching Agent <devnull@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pypandoc-binary = ">=1.13"
weasyprint = ">=58.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"

[tool.poetry.scripts]
md2pdf = "simple_converter:main"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
